# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/currencies:
    get:
      tags:
        - Currencies
      summary: List Currencies
      description: Get a paginated list of Currencies with optional filtering and sorting
      operationId: listCurrencies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Currencies
      summary: Create Currency
      description: Create a new Currency
      operationId: createCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/currency_exchanges:
    get:
      tags:
        - CurrencyExchanges
      summary: List CurrencyExchanges
      description: Get a paginated list of CurrencyExchanges with optional filtering and sorting
      operationId: listCurrencyExchanges
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyExchangeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CurrencyExchanges
      summary: Create CurrencyExchange
      description: Create a new CurrencyExchange
      operationId: createCurrencyExchange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyExchangeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyExchange'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Currency:
      type: object
      required:
        - id
        - iso_code
        - name
        - decimal_places
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        iso_code:
          type: string
          maxLength: 3
        name:
          type: string
          maxLength: 80
        symbol:
          nullable: true
          type: string
          maxLength: 8
        decimal_places:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CreateCurrencyRequest:
      type: object
      required:
        - iso_code
        - name
        - decimal_places
        - status
        - metadata
      properties:
        iso_code:
          type: string
          maxLength: 3
        name:
          type: string
          maxLength: 80
        symbol:
          nullable: true
          type: string
          maxLength: 8
        decimal_places:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CurrencyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Currency'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CurrencyExchange:
      type: object
      required:
        - id
        - from_currency
        - to_currency
        - rate
        - effective_from
        - rate_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        from_currency:
          type: string
          maxLength: 3
        to_currency:
          type: string
          maxLength: 3
        rate:
          type: string
        effective_from:
          type: string
          format: date
        effective_to:
          nullable: true
          type: string
          format: date
        rate_type:
          type: string
        source:
          nullable: true
          type: string
          maxLength: 60
        metadata:
          type: object
...
    CreateCurrencyExchangeRequest:
      type: object
      required:
        - from_currency
        - to_currency
        - rate
        - effective_from
        - rate_type
        - metadata
      properties:
        company_id:
          nullable: true
          type: string
          format: uuid
        from_currency:
          type: string
          maxLength: 3
        to_currency:
          type: string
          maxLength: 3
        rate:
          type: string
        effective_from:
          type: string
          format: date
        effective_to:
          nullable: true
          type: string
          format: date
        rate_type:
          type: string
        source:
          nullable: true
          type: string
          maxLength: 60
        metadata:
          type: object
...
    CurrencyExchangeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyExchange'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []